pyproject.toml
[tool.poetry]
name = "debts"
version = "1.0.0"
description = "The debts management service"
authors = ["Evgeniy Afonasev <ea.afonasev@gmail.com>"]
license = "MIT"
readme = 'README.md'
repository = "https://github.com/afonasev/debts"
homepage = "https://pypi.org/project/debts"
keywords = []

[tool.poetry.dependencies]
python = "^3.7"
fastapi = {version = ">=0.18,<=2.0", extras = ["all"]}
passlib = {version = "^1.7", extras = ["bcrypt"]}
psycopg2 = "^2.8"
pyjwt = "^1.7"
sqlalchemy = "^1.3"

[tool.poetry.dev-dependencies]
black = {version = "^19.2", allows-prereleases = true}
factory_boy = "^2.12"
flake8-awesome = "^1.0"
flake8-pytest-style = "^0.1.2"
mypy = "^0.660"
pylint = "^2.2"
pytest = "^3.0"
pytest-cov = "^2.6"
pytest-deadfixtures = "^2.1"
pytest-mock = "^1.10"
unify = "^0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
